% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterNode.R
\name{masterNode}
\alias{masterNode}
\title{The Master Node}
\usage{
masterNode(
  lst.running.nodes,
  operative_folder.dir,
  masterDescriptionSkel,
  output_folder,
  masterScript = c(),
  nodeScript = c()
)
}
\arguments{
\item{parameters.list}{a list containing the parameters. The possible ones are: 'considerAutoLoop' and 'threshold'. 'considerAutoLoop' is a boolean which indicates if the autoloops have to be admitted, while 'threshold' is the minimum value that a probability should have to do not be set to zero, in the transition matrix.}
}
\description{
The Master Node
}
